#19 added mock back end service for server side pagination
